Move tasks into init for now
